Improve user details validation
The validation functions were home-made ones and pretty basic.
Fortunately shinyvalidate offer and function that can validate
the content of inputs and also provide a visual feedback to users
so they can fixed the problem. Also a reactive values properly
setup was added to handle the user consent in a better way.